+++ packages/nextjs/package.json
@@ -23,6 +23,7 @@
 		"postcss": "^8.4.14",
 		"react": "18.2.0",
 		"react-dom": "18.2.0",
+		"react-loading-skeleton": "^3.3.1",
 		"siwe": "^2.1.4",
 		"tailwindcss": "^3.0.24",
 		"viem": "^1.5.0",
