+++ packages/nextjs/src/components/games/Board.tsx
@@ -83,6 +83,7 @@ function renderPockets(tmp, board, gameID) {
 	}
 	return pockets
 }
+
 const Join = gameID => {
 	const { config, refetch } = usePrepareContractWrite({
 		address: '0x98954ff59b91da3F183e9BA0111A25Be7778B7C0',
