+++ packages/nextjs/src/components/games/Kalah.tsx
