+++ packages/nextjs/src/pages/games/[gameid].tsx
@@ -1,7 +1,7 @@
-import React from 'react'
-import Kalah from '@/components/Kalah'
+import Kalah from '@/components/games/Kalah'
 import { useRouter } from 'next/router'
 import Header from '@/components/header/Header'
+import Skeleton from 'react-loading-skeleton'
 
 const Game = (): JSX.Element => {
 	const router = useRouter()
@@ -11,7 +11,7 @@ const Game = (): JSX.Element => {
 			<div className=" bg-light dark:bg-dark w-full sm:items-center py-4 sm:pt-0">
 				<Header />
 				<div className="flex flex-col items-center justify-center min-h-screen py-2">
-					<h1>Loading...</h1>
+					<Skeleton className="rounded-2xl" width={944} height={272} />
 				</div>
 			</div>
 		)
